server/pyproject.toml
[project]
name = "waste-mgt-backend"
version = "0.1.0"
description = "AI-powered waste education and recycling platform."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.14",
    "firecrawl>=2.14.0",
    "firecrawl-py>=2.14.0",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.0.10",
    "langgraph>=0.5.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
